+++ Samples/CalculatorConsole/CalculatorConsole/Operations/DivideOperation.cs
@@ -3,6 +3,6 @@
     [Operation("/", Name = "Division")]
     public class DivideOperation : IOperation
     {
-        public int Compute(int op1, int op2) => op1 / op2;
+        public double Compute(double op1, double op2) => op1 / op2;
     }
 }
